+++ precious-core/src/paths/matcher.rs
@@ -1,69 +1,67 @@
 use anyhow::Result;
-use globset::{Glob, GlobSet, GlobSetBuilder};
+use ignore::gitignore::{Gitignore, GitignoreBuilder};
 use std::path::Path;
 
 #[derive(Debug)]
 pub struct MatcherBuilder {
-    builder: GlobSetBuilder,
+    builder: GitignoreBuilder,
 }
 
 #[allow(clippy::new_without_default)]
 impl MatcherBuilder {
-    pub fn new() -> Self {
+    pub fn new<P: AsRef<Path>>(root: P) -> Self {
         Self {
-            builder: GlobSetBuilder::new(),
+            builder: GitignoreBuilder::new(root),
         }
     }
 
     pub fn with(mut self, globs: &[impl AsRef<str>]) -> Result<Self> {
         for g in globs {
-            self.builder.add(Glob::new(g.as_ref())?);
+            self.builder.add_line(None, g.as_ref())?;
         }
         Ok(self)
     }
 
     pub fn build(self) -> Result<Matcher> {
         Ok(Matcher {
-            globs: self.builder.build()?,
+            gitignore: self.builder.build()?,
         })
     }
 }
 
 #[derive(Debug)]
 pub struct Matcher {
-    globs: GlobSet,
+    gitignore: Gitignore,
 }
 
 impl Matcher {
-    pub fn path_matches(&self, path: &Path) -> bool {
-        self.globs.is_match(path)
+    pub fn path_matches(&self, path: &Path, is_dir: bool) -> bool {
+        self.gitignore.matched(path, is_dir).is_ignore()
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use anyhow::Result;
     use serial_test::parallel;
-    use std::path::PathBuf;
-
-    struct TestSet {
-        globs: Vec<String>,
-        yes: &'static [&'static str],
-        no: &'static [&'static str],
-    }
 
     #[test]
     #[parallel]
     fn path_matches() -> Result<()> {
-        let tests = vec![
+        struct TestSet {
+            globs: &'static [&'static str],
+            yes: &'static [&'static str],
+            no: &'static [&'static str],
+        }
+
+        let tests = &[
             TestSet {
-                globs: vec![String::from("*.foo")],
+                globs: &["*.foo"],
                 yes: &["file.foo", "./file.foo"],
                 no: &["file.bar", "./file.bar"],
             },
             TestSet {
-                globs: vec![String::from("*.foo"), String::from("**/foo/*")],
+                globs: &["*.foo", "**/foo/*"],
                 yes: &[
                     "file.foo",
                     "/baz/bar/file.foo",
@@ -80,18 +78,35 @@ mod tests {
                 ],
             },
             TestSet {
-                globs: vec![String::from("/foo/**/*")],
+                globs: &["/foo/**/*"],
                 yes: &["/foo/file.go", "/foo/bar/baz/file.go"],
                 no: &["/bar/file.go"],
             },
+            TestSet {
+                globs: &["/foo/**/*", "!/foo/bar/baz.*"],
+                yes: &["/foo/file.go", "/foo/bar/quux/file.go"],
+                no: &["/bar/file.go", "/foo/bar/baz.txt"],
+            },
         ];
+
         for t in tests {
-            let m = MatcherBuilder::new().with(&t.globs)?.build()?;
+            let globs = t.globs.join(" ");
+            let m = MatcherBuilder::new("/").with(t.globs)?.build()?;
             for y in t.yes {
-                assert!(m.path_matches(&PathBuf::from(y)), "{} matches", y);
+                assert!(
+                    m.path_matches(Path::new(y), false),
+                    "{} matches [{}]",
+                    y,
+                    globs,
+                );
             }
             for n in t.no {
-                assert!(!m.path_matches(&PathBuf::from(n)), "{} matches", n);
+                assert!(
+                    !m.path_matches(Path::new(n), false),
+                    "{} does not match [{}]",
+                    n,
+                    globs,
+                );
             }
         }